make sure find in files doesn't show in explorer of agents app
