+++ src/vs/workbench/contrib/search/browser/searchActionsFind.ts
@@ -12,7 +12,6 @@ import { ViewContainerLocation } from '../../../common/views.js';
 import { IViewsService } from '../../../services/views/common/viewsService.js';
 import * as Constants from '../common/constants.js';
 import * as SearchEditorConstants from '../../searchEditor/browser/constants.js';
-import { OpenSearchEditorArgs } from '../../searchEditor/browser/searchEditor.contribution.js';
 import { ISearchConfiguration, ISearchConfigurationProperties } from '../../../services/search/common/search.js';
 import { URI } from '../../../../base/common/uri.js';
 import { IsSessionsWindowContext } from '../../../common/contextkeys.js';
@@ -28,31 +27,11 @@ import { ExplorerFolderContext, ExplorerRootContext, FilesExplorerFocusCondition
 import { IPaneCompositePartService } from '../../../services/panecomposite/browser/panecomposite.js';
 import { ExplorerViewPaneContainer } from '../../files/browser/explorerViewlet.js';
 import { onUnexpectedError } from '../../../../base/common/errors.js';
-import { category, getElementsToOperateOn, getSearchView, openSearchView } from './searchActionsBase.js';
-import { IConfigurationResolverService } from '../../../services/configurationResolver/common/configurationResolver.js';
-import { IHistoryService } from '../../../services/history/common/history.js';
-import { Schemas } from '../../../../base/common/network.js';
+import { category, findInFilesCommand, getElementsToOperateOn, getSearchView, IFindInFilesArgs, openSearchView } from './searchActionsBase.js';
 import { IEditorGroupsService } from '../../../services/editor/common/editorGroupsService.js';
 import { IEditorService } from '../../../services/editor/common/editorService.js';
 import { forcedExpandRecursively } from './searchActionsTopBar.js';
 import { RenderableMatch, ISearchTreeFileMatch, ISearchTreeFolderMatchWithResource, ISearchResult, isSearchTreeFileMatch, isSearchTreeMatch } from './searchTreeModel/searchTreeCommon.js';
-
-//#region Interfaces
-export interface IFindInFilesArgs {
-	query?: string;
-	replace?: string;
-	preserveCase?: boolean;
-	triggerSearch?: boolean;
-	filesToInclude?: string;
-	filesToExclude?: string;
-	isRegex?: boolean;
-	isCaseSensitive?: boolean;
-	matchWholeWord?: boolean;
-	useExcludeSettingsAndIgnoreFiles?: boolean;
-	onlyOpenEditors?: boolean;
-	showIncludesExcludes?: boolean;
-}
-//#endregion
 
 registerAction2(class RestrictSearchToFolderAction extends Action2 {
 	constructor() {
@@ -438,67 +417,6 @@ function getMultiSelectedSearchResources(viewer: WorkbenchCompressibleAsyncDataT
 	return getElementsToOperateOn(viewer, currElement, sortConfig)
 		.map((renderableMatch) => ((isSearchTreeMatch(renderableMatch)) ? null : renderableMatch.resource))
 		.filter((renderableMatch): renderableMatch is URI => (renderableMatch !== null));
-}
-
-export async function findInFilesCommand(accessor: ServicesAccessor, _args: IFindInFilesArgs = {}) {
-
-	const searchConfig = accessor.get(IConfigurationService).getValue<ISearchConfiguration>().search;
-	const viewsService = accessor.get(IViewsService);
-	const commandService = accessor.get(ICommandService);
-	const args: IFindInFilesArgs = {};
-	if (Object.keys(_args).length !== 0) {
-		// resolve variables in the same way as in
-		// https://github.com/microsoft/vscode/blob/8b76efe9d317d50cb5b57a7658e09ce6ebffaf36/src/vs/workbench/contrib/searchEditor/browser/searchEditorActions.ts#L152-L158
-		const configurationResolverService = accessor.get(IConfigurationResolverService);
-		const historyService = accessor.get(IHistoryService);
-		const workspaceContextService = accessor.get(IWorkspaceContextService);
-		const activeWorkspaceRootUri = historyService.getLastActiveWorkspaceRoot();
-		const filteredActiveWorkspaceRootUri = activeWorkspaceRootUri?.scheme === Schemas.file || activeWorkspaceRootUri?.scheme === Schemas.vscodeRemote ? activeWorkspaceRootUri : undefined;
-		const lastActiveWorkspaceRoot = filteredActiveWorkspaceRootUri ? workspaceContextService.getWorkspaceFolder(filteredActiveWorkspaceRootUri) ?? undefined : undefined;
-
-		for (const entry of Object.entries(_args)) {
-			const name = entry[0];
-			const value = entry[1];
-			if (value !== undefined) {
-				// eslint-disable-next-line local/code-no-any-casts
-				(args as any)[name as any] = (typeof value === 'string') ? await configurationResolverService.resolveAsync(lastActiveWorkspaceRoot, value) : value;
-			}
-		}
-	}
-
-	const mode = searchConfig?.mode;
-	if (mode === 'view') {
-		openSearchView(viewsService, false).then(openedView => {
-			if (openedView) {
-				const searchAndReplaceWidget = openedView.searchAndReplaceWidget;
-				searchAndReplaceWidget.toggleReplace(typeof args.replace === 'string');
-				let updatedText = false;
-				if (typeof args.query !== 'string') {
-					updatedText = openedView.updateTextFromFindWidgetOrSelection({ allowUnselectedWord: typeof args.replace !== 'string' });
-				}
-				openedView.setSearchParameters(args);
-				if (typeof args.showIncludesExcludes === 'boolean') {
-					openedView.toggleQueryDetails(false, args.showIncludesExcludes);
-				}
-
-				openedView.searchAndReplaceWidget.focus(undefined, updatedText, updatedText);
-			}
-		});
-	} else {
-		const convertArgs = (args: IFindInFilesArgs): OpenSearchEditorArgs => ({
-			location: mode === 'newEditor' ? 'new' : 'reuse',
-			query: args.query,
-			filesToInclude: args.filesToInclude,
-			filesToExclude: args.filesToExclude,
-			matchWholeWord: args.matchWholeWord,
-			isCaseSensitive: args.isCaseSensitive,
-			isRegexp: args.isRegex,
-			useExcludeSettingsAndIgnoreFiles: args.useExcludeSettingsAndIgnoreFiles,
-			onlyOpenEditors: args.onlyOpenEditors,
-			showIncludesExcludes: !!(args.filesToExclude || args.filesToExclude || !args.useExcludeSettingsAndIgnoreFiles),
-		});
-		commandService.executeCommand(SearchEditorConstants.OpenEditorCommandId, convertArgs(args));
-	}
 }
 
 async function modifySearchFileTypePattern(accessor: ServicesAccessor, fileMatch: ISearchTreeFileMatch | undefined, isExclude: boolean) {
